broke out flights page into subcomponents
- now the flights page has more subcomponents, mroe modular, etc...
